requirements.txt
streamlit==1.32.0
pandas==2.3.2
google-generativeai==0.3.1
python-dotenv==1.0.0
numpy==1.24.3
matplotlib==3.7.1